Move more things out of index
